All bizniz logicks transfer to the action modo
