+++ SurfsUpAPI/Models/Surfboard.cs
@@ -6,7 +6,7 @@
         public string BoardName { get; set; }
         public float Length { get; set; }
         public float Width { get; set; }
-        public float Thickness { get; set; }
+        public float Thickness {  get; set; }
         public float Volume { get; set; }
 
         public int Price { get; set; } // måske vi skal vælge en anden vartype
@@ -18,7 +18,7 @@
         //     None, Shortboard, Funboard, Fish, Longboard, SUP
         // }
 
-        public Surfboard(int surfboardId, string boardName, float length, float width, float thickness, float volume, /*BoardType boardtype,*/ int price, string equipment)
+        public Surfboard (int surfboardId, string boardName, float length, float width, float thickness, float volume, /*BoardType boardtype,*/ int price, string equipment)
         {
             SurfboardId = surfboardId;
             BoardName = boardName;
@@ -33,5 +33,8 @@
         public Surfboard()
         {
         }
+
+
     }
+    
 }
