+++ SurfsUpv3/Views/Admins/Create.cshtml
@@ -1,4 +1,4 @@
-﻿@model SurfsUpv3.Models.Booking
+﻿@model SurfsUpAPI.Models.Booking
 
 @{
     ViewData["Title"] = "Create";
