+++ SurfsUpv3/Views/Admins/Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@model SurfsUpv3.Models.Booking
+﻿@model SurfsUpAPI.Models.Booking
 
 @{
     ViewData["Title"] = "Edit";
