+++ SurfsUpv3/Views/Admins/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<SurfsUpv3.Models.Booking>
+﻿@model IEnumerable<SurfsUpAPI.Models.Booking>
 
 @{
     ViewData["Title"] = "Index";
