+++ SurfsUpv3/Views/RentOrder/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@using SurfsUpv3.Models;
+﻿@using SurfsUpAPI.Models;
 @model Booking
