+++ SurfsUpv3/Views/Surfboards/Create.cshtml
@@ -1,4 +1,4 @@
-﻿@model SurfsUpv3.Models.Surfboard
+﻿@model SurfsUpAPI.Models.Surfboard
 
 @{
     ViewData["Title"] = "Create";
