+++ SurfsUpv3/Views/Surfboards/Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@model SurfsUpv3.Models.Surfboard
+﻿@model SurfsUpAPI.Models.Surfboard
 
 @{
     ViewData["Title"] = "Edit";
