+++ SurfsUpv3/Views/Surfboards/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<SurfsUpv3.Models.Surfboard>
+﻿@model IEnumerable<SurfsUpAPI.Models.Surfboard>
 
 @{
     ViewData["Title"] = "Index";
