+++ SurfsUpv3/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
 ﻿@using SurfsUpv3
-@using SurfsUpv3.Models
+@using SurfsUpAPI.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
